==> scripts/02_qc/02_quality_control_post_cellranger.Rmd <==
---
title: "10X QC processing post cellranger"
author: "Kimberly Olney, PhD"
date: "03/11/2025"
output:
  html_document:
    theme: cerulean
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---
# Lewy body dementia center without walls LBD CWOW

# Set working directory
```{r working_directory}
knitr::opts_knit$set(root.dir = "/tgen_labs/jfryer/kolney/LBD_CWOW/LBD_snRNA/scripts/")
setwd("/tgen_labs/jfryer/kolney/LBD_CWOW/LBD_snRNA/scripts/")
```

# Libraris, paths, colors
```{r echo=FALSE, message=FALSE}
source(here::here("/tgen_labs/jfryer/kolney/LBD_CWOW/LBD_snRNA/scripts/", "file_paths_and_colours.R"))
projectID <- "CWOW_cellranger"

# Remove sample
metadata <- subset(metadata, Sample_ID != "LBD_AS_F4")
metadata$sampleID <- factor(metadata$Sample_ID, levels = c(metadata$Sample_ID))
samples <- metadata$sampleID 
sex_order <- factor(metadata$sex_inferred, levels = unique(metadata$sex_inferred))
disease_order <- factor(metadata$TYPE, levels = c("CONTROL", "AD_AT", "LBD_S", "LBD_AS", "LBD_ATS"))

metadata <- metadata %>%
  mutate(sampleID = gsub(".*_(\\d+)_.*_(BR_Nuclei).*", "\\2_\\1", Lane.Name))
samples <- metadata$sampleID 

# sampleID with disease_order
order <- metadata %>%
  arrange(disease_order) %>%
  dplyr::select(TYPE, sampleID, Sample_ID)
write.table(order, "order.txt", quote = F, row.names = F, sep = "\t")
samples <- order$sampleID
disease_order <- order$TYPE
sample_order <- factor(order$Sample_ID, levels = order$Sample_ID)
```

# Genes 
```{r gene_info}
if (file.exists("../../rObjects/annotation.rds")) {
  genes <- readRDS("../../rObjects/annotation.rds")
} else {
  gtf.file <- paste0(pathToRef, "/genes/genes.gtf")
  genes <- rtracklayer::import(gtf.file)
  genes <- as.data.frame(genes)
  genes <- genes[genes$type == "gene",]
  saveRDS(genes, "../rObjects/annotation.rds")
}

gene_type_table <- table(genes$gene_type)
write.table(gene_type_table, "gene_type_table.tsv", row.names = F, quote = F, sep = "\t")
mt.genes.df <- subset(genes, seqnames == "chrM")
mt.genes <- mt.genes.df$gene_name
```

# QC thresholds
nCount_RNA = total number of transcripts (UMIs) in a single cell 
nFeature_RNA = number of unique genes (features)
```{r QC_metrics}
nCount.min <- 1000 
nCount.max <- 100000 
nFeature.min <- 800
nFeature.max<- 12000
complexity.cutoff <- 0.80
mt.cutoff <- 5
hb.cutoff <- 1
ribo.cutoff <- 1
choroid.cutoff <- 1
```

# Setup seurat object
```{r seurat_object}
dataObject <- readRDS(paste0("../../rObjects/", projectID, ".rds"))
```

# Change indents
```{r change_idents}
barcodes <- colnames(dataObject)
sample <- str_match(barcodes, "(.+)_[ACGT]+")[,2]
table(sample)
unique(sample)
dataObject$sample <- factor(sample, levels = samples)
table(dataObject$sample)  # check
Idents(dataObject) <- dataObject$sample
```

Add metadata to seurate object
```{r condition_column}
# Match the samples to the dataObject
#sample_match <- match(dataObject$sample, samples)
seurat_sample_order <- as.character(dataObject$sample)
matched_metadata <- metadata[match(seurat_sample_order, as.character(metadata$sampleID)), ]

# Verify the order after matching
if (!all.equal(seurat_sample_order, matched_metadata$sampleID)) {
  stop("Sample order mismatch after matching metadata. Please check your sample IDs.")
}

# Add the metadata columns to the Seurat object's metadata
dataObject$Sample_ID <- factor(matched_metadata$Sample_ID, levels = unique(matched_metadata$Sample_ID))
dataObject$sex <- matched_metadata$sex_inferred
dataObject$age <- matched_metadata$Age
dataObject$PathDx <- matched_metadata$PathDx
dataObject$group <- matched_metadata$TYPE
dataObject$group <- factor(dataObject$group, levels = c("CONTROL", "AD_AT", "LBD_S", "LBD_AS", "LBD_ATS"))
matched_metadata$Cing.LB[is.na(matched_metadata$Cing.LB)] <- 0
matched_metadata$Braak.NFT[is.na(matched_metadata$Braak.NFT)] <- 0
matched_metadata$Thal.amyloid[is.na(matched_metadata$Thal.amyloid)] <- 0
matched_metadata$MF.SP[is.na(matched_metadata$MF.SP)] <- 0
matched_metadata$MF.NFT[is.na(matched_metadata$MF.NFT)] <- 0
matched_metadata$MF.LB[is.na(matched_metadata$MF.LB)] <- 0
matched_metadata$MF.Amyloid[is.na(matched_metadata$MF.Amyloid)] <- 0
matched_metadata$MF.Tau[is.na(matched_metadata$MF.Tau)] <- 0
matched_metadata$Cing.Synuclein[is.na(matched_metadata$Cing.Synuclein)] <- 0
dataObject$ClinicalDx <- matched_metadata$ClinicalDx
dataObject$FHx <- matched_metadata$FHx
dataObject$APOE <- matched_metadata$APOE
dataObject$MAPT <- matched_metadata$MAPT
dataObject$GRN <- matched_metadata$GRN
dataObject$TMEM106b <- matched_metadata$TMEM106b
dataObject$RIN <- matched_metadata$RIN
```

nuclei count per sample, sex, and group
```{r}
# Sample
table(dataObject$sample) 
table(dataObject$Sample_ID) 

# Sex 
table(dataObject$sex) 

# Type
table(dataObject$group) 
```


QC columns 
```{r QC_columns}
summary(dataObject$nCount_RNA)
summary(dataObject$nFeature_RNA)
# cell.complexity
dataObject$cell.complexity <- log10(dataObject$nFeature_RNA) / log10(dataObject$nCount_RNA)

# Chromosome M
gene.names <- rownames(dataObject)
dataObject$percent.mt <- PercentageFeatureSet(dataObject, features = mt.genes)
summary(dataObject$percent.mt)

"^RPS|^RPL"
# ribosomal proteins 
ribo.genes <- gene.names[grep("^RPS|^RPL", gene.names)] 
mt.ribo <- gene.names[grep("^MRP[SL]", gene.names)]
ribo.combined <- c(mt.ribo,ribo.genes)
dataObject$percent.ribo <- PercentageFeatureSet(dataObject, features = ribo.combined)
summary(dataObject$percent.ribo)

# hemoglobin proteins
hb.genes <- gene.names[grep("^HB[BA]", gene.names)]
dataObject$percent.hb <- PercentageFeatureSet(dataObject, features = hb.genes)
summary(dataObject$percent.hb)

# percent choroid plexus
dataObject$percent.choroid <- PercentageFeatureSet(dataObject, features = c("TTR","FOLR1", "PRLR"))
summary(dataObject$percent.choroid)
```

# Pre-filtering QC
## Number of cells
```{r prefiltering_cells_per_sample}
# Visualize the number of cell counts per sample
data <- as.data.frame(table(dataObject$Sample_ID))
colnames(data) <- c("Sample_ID","frequency")

ncells1 <- ggplot(data, aes(x = Sample_ID, y = frequency, fill = Sample_ID)) + 
  geom_col() +
  theme_classic() +
  geom_text(aes(label = frequency), 
            position=position_dodge(width=0.9), 
            #vjust=-0.25, 
            hjust = -.025,
            angle = 90) +
  #scale_fill_manual(values = sample_colors) + 
  scale_y_continuous(breaks = seq(0,30000, by = 2000), limits = c(0,30000)) +
  ggtitle("Raw: nuclei per sample") +
  theme(legend.position =  "none") + 
  theme(axis.text.x = element_text(angle = 45, hjust=1))
ncells1
```

## Density plots
```{r prefiltering_density}
# set graphical parameter
par(mfrow = c(4,1))
# Visualize nCount_RNA
den1 <- ggplot(dataObject@meta.data,
       aes(color = Sample_ID,
           x = nCount_RNA,
           fill = Sample_ID)) +
  geom_density(alpha = 0.2) +
  theme_classic() +
  scale_x_log10() +
  #scale_color_manual(values = sample_colors) +
  #scale_fill_manual(values = sample_colors) +
  xlab("nCount_RNA") +
  ylab("Density") +
  geom_vline(xintercept = nCount.min) +
  geom_vline(xintercept = nCount.max) + theme(legend.position="none")


# Visualize nFeature
den2 <- ggplot(dataObject@meta.data,
       aes(color = Sample_ID,
           x = nFeature_RNA,
           fill = Sample_ID)) +
  geom_density(alpha = 0.2) +
  theme_classic() +
  scale_x_log10() +
  #scale_color_manual(values = sample_colors) +
  #scale_fill_manual(values = sample_colors) +
  xlab("nFeature_RNA") +
  ylab("Density") +
  geom_vline(xintercept = nFeature.min) +
  geom_vline(xintercept = nFeature.max) + theme(legend.position="none")



# Visualize cell complexity
# Quality cells are usually above 0.85
den3 <- ggplot(dataObject@meta.data,
       aes(color = Sample_ID,
           x = cell.complexity,
           fill = Sample_ID)) +
  geom_density(alpha = 0.2) +
  theme_classic() +
  #scale_color_manual(values = sample_colors) +
  #scale_fill_manual(values = sample_colors) +
  xlab("Cell Complexity (log10(nFeature/nCount))") +
  ylab("Density") +
  geom_vline(xintercept = complexity.cutoff) + theme(legend.position="none")

# Visualize percent.mt
den4 <- ggplot(dataObject@meta.data,
       aes(color = Sample_ID,
           x = percent.mt,
           fill = Sample_ID)) +
  geom_density(alpha = 0.2) +
  theme_classic() +
  scale_x_continuous(n.breaks = 4) +
  geom_vline(xintercept = mt.cutoff) +
  #scale_color_manual(values = sample_colors) +
  #scale_fill_manual(values = sample_colors) +
  xlab("% Mitochondrial Genes") +
  ylab("Density")+ theme(legend.position="none")



# Arrange graphs in grid
plots1 <- list(den1,den2,den3,den4)
layout1 <- rbind(c(1),c(2),c(3),c(4))
grid1 <- grid.arrange(grobs = plots1, layout_matrix = layout1)
```

## Violin plots
Warning: Default search for "data" layer in "RNA" assay yielded no results; utilizing "counts" layer instead."

This warning tells you that Seurat, by default, prefers to plot from the data (normalized) slot. However, we haven't normalized the RNA assay, thus the data slot is empty. As a fallback, Seurat  switched to using the counts slot, which contains the raw UMI counts. 
```{r prefiltering_violins}
# nFeature, nCount, and cell.complexity violins
v1 <- VlnPlot(dataObject,
              features = c( "nCount_RNA","nFeature_RNA", "cell.complexity"),
              ncol = 1,
              group.by = 'Sample_ID',
              ##cols = sample_colors,
              pt.size = 0)
v1

#  percent violins
v2 <- VlnPlot(dataObject,
              features = c("percent.mt","percent.ribo","percent.hb", "percent.choroid"),
              ncol = 2,
              group.by = 'Sample_ID',
              #cols = sample_colors,
              pt.size = 0)
v2
```

```{r, echo=FALSE}
# save v1
v1
path <- paste0("../../results/violin/",projectID,"_nFeature_nCount_complexity_raw")
saveToPDF(paste0(path, ".pdf"), width = 12, height = 9)
dev.off()

# save v2
v2
path <- paste0("../../results/violin/",projectID,"_percent_raw")
saveToPDF(paste0(path, ".pdf"), width = 16, height = 7)
dev.off()

# cleanup
#remove(v1,v2)
```

## Scatter plots
```{r prefiltering_scatter1, warning=FALSE}
s1 <- ggplot(
  dataObject@meta.data,
  aes(x = nCount_RNA, y = nFeature_RNA, color = percent.mt)) + 
  geom_point() + 
  stat_smooth(method=lm) +
	scale_x_log10() +   	
  scale_y_log10() + 
  theme_classic() +
  geom_vline(xintercept = nCount.min) + 
  geom_vline(xintercept = nCount.max) + 
  geom_hline(yintercept = nFeature.min) + 
  geom_hline(yintercept = nFeature.max) + 
  facet_wrap(~Sample_ID, ncol = 8) +
  scale_colour_gradient(low = "gray90", high = "black", limits =c(0,100))
s1
```

```{r echo=FALSE}
# save
s1
path <- paste0("../../results/scatter/",projectID,"_nFeature_vs_nCount_percentMt_raw")
saveToPDF(paste0(path, ".pdf"), width = 20, height = 15)
dev.off()

# cleanup
remove(s1)
```

```{r prefiltering_scatter2}
s2 <- FeatureScatter(dataObject,
               feature1 = "nCount_RNA",
               feature2 = "percent.mt",
               group.by = 'Sample_ID',
               #cols = sample_colors,
               shuffle = TRUE)
s2
```

```{r,echo=FALSE}
# save
s2
path <- paste0("../../results/scatter/",projectID,"_nCount_vs_percentMT_raw")
saveToPDF(paste0(path, ".pdf"), width = 6, height = 4)

# cleanup
remove(s2)
```

# Filtering
## Cell-level filtering
We want to be careful filtering because removing things can easily lead to
misinterpretation. 
```{r}
# 1. Initial number of nuclei
initial_nuclei <- ncol(dataObject)
message(paste0("Initial number of nuclei: ", initial_nuclei))

# Create a working copy of the object
filtered_data <- dataObject

# Define your thresholds (ensure these are set based on your VlnPlot analysis)
# Already defined at the top of the script for easy referencing. 

# --- Apply Filters Sequentially and Report ---
# Filter 1: nCount_RNA
prev_nuclei <- ncol(filtered_data)
filtered_data <- subset(filtered_data, subset = nCount_RNA > nCount.min & nCount_RNA < nCount.max)
removed_count <- prev_nuclei - ncol(filtered_data)
message(paste0("Removed ", removed_count, " nuclei based on nCount_RNA (new total: ", ncol(filtered_data), ")"))

# Filter 2: nFeature_RNA
prev_nuclei <- ncol(filtered_data)
filtered_data <- subset(filtered_data, subset = nFeature_RNA > nFeature.min & nFeature_RNA < nFeature.max)
removed_count <- prev_nuclei - ncol(filtered_data)
message(paste0("Removed ", removed_count, " nuclei based on nFeature_RNA (new total: ", ncol(filtered_data), ")"))

# Filter 3: cell.complexity (assuming this metric is pre-calculated in your object)
prev_nuclei <- ncol(filtered_data)
filtered_data <- subset(filtered_data, subset = cell.complexity > complexity.cutoff)
removed_count <- prev_nuclei - ncol(filtered_data)
message(paste0("Removed ", removed_count, " nuclei based on cell.complexity (new total: ", ncol(filtered_data), ")"))

# Filter 4: percent.mt
prev_nuclei <- ncol(filtered_data)
filtered_data <- subset(filtered_data, subset = percent.mt <= mt.cutoff)
removed_count <- prev_nuclei - ncol(filtered_data)
message(paste0("Removed ", removed_count, " nuclei based on percent.mt (new total: ", ncol(filtered_data), ")"))

# Filter 5: percent.hb
prev_nuclei <- ncol(filtered_data)
filtered_data <- subset(filtered_data, subset = percent.hb <= hb.cutoff)
removed_count <- prev_nuclei - ncol(filtered_data)
message(paste0("Removed ", removed_count, " nuclei based on percent.hb (new total: ", ncol(filtered_data), ")"))

# Filter 6: percent.ribo
prev_nuclei <- ncol(filtered_data)
filtered_data <- subset(filtered_data, subset = percent.ribo <= ribo.cutoff)
removed_count <- prev_nuclei - ncol(filtered_data)
message(paste0("Removed ", removed_count, " nuclei based on percent.ribo (new total: ", ncol(filtered_data), ")"))

# Filter 7: percent.choroid
prev_nuclei <- ncol(filtered_data)
filtered_data <- subset(filtered_data, subset = percent.choroid <= choroid.cutoff)
removed_count <- prev_nuclei - ncol(filtered_data)
message(paste0("Removed ", removed_count, " nuclei based on percent.choroid (new total: ", ncol(filtered_data), ")"))


# Final filtered object
dataObject.filtered <- filtered_data
final_nuclei <- ncol(dataObject.filtered)
message(paste0("\nTotal nuclei remaining after all filters: ", final_nuclei))
message(paste0("Total nuclei removed: ", initial_nuclei - final_nuclei))


# Compare counts per Sample_ID before and after
message("\nNuclei count per Sample_ID (Original):")
print(table(dataObject$Sample_ID))
message("\nNuclei count per Sample_ID (Filtered):")
print(table(dataObject.filtered$Sample_ID))
```


```{r cell_filtering, eval = FALSE}
# filter
dataObject.filtered <- subset(dataObject,
                        subset = (nCount_RNA > nCount.min)&
                          (nCount_RNA < nCount.max) & 
                          (nFeature_RNA > nFeature.min) & 
                          (nFeature_RNA < nFeature.max) & 
                          (cell.complexity > complexity.cutoff) &
                          (percent.mt <= mt.cutoff) & 
                          (percent.hb <= hb.cutoff) &
                          (percent.ribo <= ribo.cutoff) &
                          (percent.choroid <= choroid.cutoff))

# print nuclei removed
print(paste0(dim(dataObject)[2] - dim(dataObject.filtered)[2]," nuclei removed"))

table(dataObject$Sample_ID)
table(dataObject.filtered$Sample_ID)
```

## Gene-level filtering
Remove lowly expressed genes.  We will keep genes that have at least 1 count in 10 cells.
```{r gene_filtering}
dataObject.filtered <- JoinLayers(dataObject.filtered)
counts <- GetAssayData(object = dataObject.filtered, layer = "counts")
nonzero <- counts > 0 
keep <- Matrix::rowSums(nonzero) >= 10  
counts.filtered <- counts[keep,]  

dataObject.filtered <- CreateSeuratObject(counts.filtered, 
                                     meta.data = dataObject.filtered@meta.data)

# print features removed
print(paste0(dim(counts)[1] - dim(counts.filtered)[1], " features removed"))
```

## Mitochondrial gene filtering
```{r}
# remove mt.genes
counts <- GetAssayData(object = dataObject.filtered, layer = "counts")
keep <- !rownames(counts) %in% mt.genes # false when mt.gene
counts.filtered <- counts[keep,]

# overwrite dataObject.filtered
dataObject.filtered <- CreateSeuratObject(counts.filtered, 
                                    meta.data = dataObject.filtered@meta.data)

# print features removed
print(paste0(dim(counts)[1] - dim(counts.filtered)[1], " features removed"))
```

## Save filtering
```{r,echo=FALSE}
saveRDS(dataObject.filtered, paste0("../../rObjects/",projectID,"_filtered.rds"))
#dataObject.filtered <- readRDS(paste0("../../rObjects/",projectID,"_filtered.rds"))
```

# Post-filtering
## Number of cells
```{r number_cells2}
# Visualize the number of cell counts per sample
data <- as.data.frame(table(dataObject.filtered$Sample_ID))
colnames(data) <- c("Sample_ID","frequency")

ncells2 <- ggplot(data, aes(x = Sample_ID, y = frequency, fill = Sample_ID)) + 
  geom_col() +
  theme_classic() +
  geom_text(aes(label = frequency), 
            position=position_dodge(width=0.9), 
            #vjust=-0.25, 
            hjust = -.025,
            angle = 90) +
  #scale_fill_manual(values = sample_colors) + 
  scale_y_continuous(breaks = seq(0,30000, by = 2000), limits = c(0,30000)) +
  ggtitle("Filtered: nuclei per sample") +
  theme(legend.position =  "none") + 
  theme(axis.text.x = element_text(angle = 45, hjust=1))
ncells2
# Arrange graphs in grid
plots2 <- list(ncells1,ncells2)
layout2 <- rbind(c(1),c(2))
grid2 <- grid.arrange(grobs = plots2, layout_matrix = layout2)

mean(data$frequency)
median(data$frequency)
```

```{r,echo=FALSE}
# save
grid2 <- grid.arrange(grobs = plots2, layout_matrix = layout2)
path <- paste0("../../results/nuclei_count/",projectID, 
               "_cells_per_sample")
saveToPDF(paste0(path, ".pdf"), width = 12, height = 8)


# cleanup
remove(ncells1,ncells2,plots2,layout2,grid2)
```

## Density plots
```{r postfiltering_density}
# set graphical parameter
par(mfrow = c(4,1))

# Visualize the number of counts per cell
den5 <- ggplot(dataObject.filtered@meta.data,
       aes(color = Sample_ID,
           x = nCount_RNA,
           fill = Sample_ID)) +
  geom_density(alpha = 0.2) +
  theme_classic() +
  scale_x_log10() +
  #scale_color_manual(values = sample_colors) +
  #scale_fill_manual(values = sample_colors) +
  xlab("nCount_RNA") +
  ylab("Density") +
  geom_vline(xintercept = nCount.min) +
  geom_vline(xintercept = nCount.max) + theme(legend.position="none")


den6 <- ggplot(dataObject.filtered@meta.data,
       aes(color = Sample_ID,
           x = nFeature_RNA,
           fill = Sample_ID)) +
  geom_density(alpha = 0.2) +
  theme_classic() +
  scale_x_log10() +
  #scale_color_manual(values = sample_colors) +
  #scale_fill_manual(values = sample_colors) +
  xlab("nFeature_RNA") +
  ylab("Density") +
  geom_vline(xintercept = nFeature.min) +
  geom_vline(xintercept = nFeature.max) + theme(legend.position="none")



# Visualize cell complexity
# Quality cells are usually above 0.80
den7 <- ggplot(dataObject.filtered@meta.data,
       aes(color = Sample_ID,
           x = cell.complexity,
           fill = Sample_ID)) +
  geom_density(alpha = 0.2) +
  theme_classic() +
  scale_x_log10() +
  #scale_color_manual(values = sample_colors) +
  #scale_fill_manual(values = sample_colors) +
  xlab("Cell Complexity (log10(nFeature/nCount))") +
  ylab("Density") +
  geom_vline(xintercept = complexity.cutoff) + theme(legend.position="none")


# Visualize percent.mt
den8 <- ggplot(dataObject.filtered@meta.data,
       aes(color = Sample_ID,
           x = percent.mt,
           fill = Sample_ID)) +
  geom_density(alpha = 0.2) +
  theme_classic() +
  scale_x_log10() +
  #scale_color_manual(values = sample_colors) +
  #scale_fill_manual(values = sample_colors) +
  xlab("% Mitochondrial Genes") +
  ylab("Density") +
  geom_vline(xintercept = mt.cutoff) + theme(legend.position="none")


# Arrange graphs in grid
plots3 <- list(den1,den2,den3,den4,den5,den6,den7,den8)
layout3 <- rbind(c(1,5),c(2,6),c(3,7), c(4,8))
grid3 <- grid.arrange(grobs = plots3, layout_matrix = layout3)
```

```{r, echo=FALSE}
# save
grid3 <- grid.arrange(grobs = plots3, layout_matrix = layout3)
path <- paste0("../../results/density/",projectID, 
               "_density")
saveToPDF(paste0(path, ".pdf"), width = 11, height = 8.5)

# cleanup
remove(den1,den2,den3,den4,den5,den6,plots3,layout3,grid3)
```

## Violin plots
```{r postfiltering_violins}
# nFeature, nCount, and cell.complexity violins
v3 <- VlnPlot(dataObject.filtered,
              features = c("nFeature_RNA", "nCount_RNA","cell.complexity"),
              ncol = 1,
              group.by = 'Sample_ID',
              #cols = sample_colors,
              pt.size = 0)
v3

#  percent violins
v4 <- VlnPlot(dataObject.filtered,
              features = c("percent.mt","percent.ribo","percent.hb", "percent.choroid"),
              ncol = 2,
              group.by = 'Sample_ID',
              #cols = sample_colors,
              pt.size = 0)
v4
```

```{r, echo=FALSE}
# save
v3
path <- paste0("../../results/violin/",projectID, 
               "_nFeature_nCount_complexity_filtered")
saveToPDF(paste0(path, ".pdf"), width = 12, height = 11)

v4
path <- paste0("../../results/violin/",projectID, 
               "_percent_filtered")
saveToPDF(paste0(path, ".pdf"), width = 14, height = 9)

# cleanup
remove(v3,v4)
```

## Scatter plots
```{r postfiltering_scatter1}
s3 <- ggplot(
  dataObject.filtered@meta.data,
  aes(x = nCount_RNA, y = nFeature_RNA, color = percent.mt)) + 
  geom_point() + 
  stat_smooth(method=lm) +
	scale_x_log10() +   	
  scale_y_log10() + 
  theme_classic() +
  geom_vline(xintercept = nCount.min) + 
  geom_vline(xintercept = nCount.max) + 
  geom_hline(yintercept = nFeature.min) + 
  geom_hline(yintercept = nFeature.max) + 
  facet_wrap(~Sample_ID, ncol = 8) +
  scale_colour_gradient(low = "gray90", high = "black", limits =c(0,100))
s3
```

```{r,echo=FALSE}
# save
s3
path <- paste0("../../results/scatter/",projectID,
               "_nFeature_vs_nCount_perecentMt_filtered")
saveToPDF(paste0(path, ".pdf"), width = 20, height = 15)

# cleanup
remove(s3)
```

```{r postfiltering_scatter2}
s4 <- FeatureScatter(dataObject.filtered,
               feature1 = "nCount_RNA",
               feature2 = "percent.mt",
               group.by = 'Sample_ID',
               #cols = sample_colors,
               shuffle = TRUE)
s4
```

```{r,echo=FALSE}
# save
s4
path <- paste0("../../results/scatter/",projectID,
               "_nCount_vs_percentMT_filtered")
saveToPDF(paste0(path, ".pdf"), width = 6, height = 4)

# cleanup
remove(s4)
```

## Box plot
```{r boxplot}
# Visualize the distribution of genes detected per cell via boxplot
b1 <- ggplot(dataObject.filtered@meta.data,
       aes(x = Sample_ID, 
           y = log10(nFeature_RNA), 
           fill=Sample_ID)) + 
  geom_boxplot() + 
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1),
        legend.position = "none") +
  theme(plot.title = element_text(hjust = 0.5, face="bold")) +
  ggtitle("Unique Genes / Nuclei / Sample") +
  #scale_color_manual(values = sample_colors) +
  #scale_fill_manual(values = sample_colors) +
  xlab("Sample")
b1
```

```{r,echo=FALSE}
# save
b1
path <- paste0("../../results/nuclei_count/",projectID,
               "_nFeature_per_sample")
saveToPDF(paste0(path, ".pdf"), width = 12, height = 4)

# cleanup
remove(b1)
```

## Top transcripts
```{r top_transcripts}
df <- data.frame(row.names = rownames(dataObject.filtered))
df$rsum <- rowSums(x = dataObject.filtered, slot = "counts")
df$gene_name <- rownames(df)
df <- df[order(df$rsum,decreasing = TRUE),]
head(df, 10)
```

```{r,echo=FALSE}
write.table(df, paste0("../../results/top_transcripts/", projectID, "_abundant_transcripts.txt"),
            quote = FALSE,
            row.names = FALSE)
```
